+++ src/models/ActionAccept.php
@@ -4,15 +4,7 @@ namespace HtmlAcademy\models;
 class ActionAccept extends Actions {
 
     public static function getName() {
-
-    }
-
-    public static function getInnerName() {
-
-    }
-
-    public static function checkAccess() {
-
+        return self::class;
     }
 
 }
