+++ src/models/ActionReject.php
@@ -4,15 +4,7 @@ namespace HtmlAcademy\models;
 class ActionReject extends Actions {
 
     public static function getName() {
-
-    }
-
-    public static function getInnerName() {
-
-    }
-
-    public static function checkAccess() {
-
+        return self::class;
     }
 
 }
